fixed the global state - added validation - added submitquestion - finnished all the apicall functions

+++ src/components/DashBoard/index.js
@@ -9,22 +9,40 @@ import Users from "./DashboardContents/Users/Users"
 import Questions from "./DashboardContents/Questions/Questions"
 
 import { useNavigate } from "react-router-dom"
+import { getUserDetails } from "../../core/api"
 
 // imports
 import { useStoreActions, useStoreState } from "easy-peasy"
 
 function Home() {
-  // const user = useStoreState(({ User }) => User.user)
-  const fetchUserDetails = useStoreActions(({ User }) => User.fetchUserDetails)
+  const actions = useStoreActions((actions) => ({
+    fetchUserDetails: actions.User.fetchUserDetails,
+    setUser: actions.User.setUser,
+    setLoading: actions.User.setLoading,
+  }))
 
   const navigate = useNavigate()
   const fetchData = async () => {
-    try {
-      await fetchUserDetails()
-    } catch (e) {
-      navigate("/Login")
-      localStorage.clear()
-    }
+    getUserDetails()
+      .then(({ data: res }) => {
+        const { status, message, data } = res
+
+        console.log(res)
+        if (status == "success") {
+          actions.setLoading(false)
+          actions.setUser(data)
+        }
+      })
+      .catch((e) => {
+        console.log(e)
+        if (e?.response) {
+          // const {data: {status, message, error}} = e.response
+          if (e.response.status == 401) {
+            navigate("/Login") // add queryString
+            localStorage.clear()
+          }
+        }
+      })
   }
 
   useEffect(() => {
